secure_banking_app/requirements.txt
Flask==2.1.1
pymongo==4.2.0
bcrypt==3.2.0
pyotp==2.5.0
PyJWT==2.4.0
requests==2.27.1
pyopenssl==22.0.0
tk
